+++ src/light.cpp
@@ -16,7 +16,7 @@ Light::Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLflo
     diffuseIntensity = dIntensity;  
 }
 
-Light::Light(GLfloat shadowWidth, GLfloat shadowHeight, GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity){
+Light::Light(GLuint shadowWidth, GLuint shadowHeight, GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity){
 
     color = glm::vec3(red, green, blue); 
     ambientIntensity = aIntensity;
