+++ src/light.h
@@ -15,7 +15,7 @@ public:
 
     Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity);
 
-    Light(GLfloat shadowWidth, GLfloat shadowHeight,
+    Light(GLuint shadowWidth, GLuint shadowHeight,
         GLfloat red, GLfloat green, GLfloat blue, 
         GLfloat aIntensity, GLfloat dIntensity);
